Rework component organization
And fix the refresh button bug
